Fixed Fluid recipe NBT write
Wow, forge overrides equals() and hashCode() in FluidStack but did not
int Fluid. It cause fluid has different hash after restart of JVM. Fixed
by adding custom hashCode in GT_Utility. Works fine. Closes #74.

+++ src/main/java/gregtechmod/common/recipe/logic/GeneratorRecipeLogic.java
@@ -85,7 +85,7 @@ public class GeneratorRecipeLogic extends RecipeLogic {
 		if (getMachine().spaceForOutput(recipe)) {
 			previousRecipe = recipe;
 			progressTime = 1;
-			leftEU = (int) (recipe.getDuration() * recipe.getEUt() * (efficiency.getAsInt() / 100.0D));
+			leftEU = (long) (recipe.getDuration() * recipe.getEUt() * (efficiency.getAsInt() / 100.0D));
 			maxProgressTime = (int) Math.ceil(leftEU * 1.0D / ((MetaTileEntity) getMachine()).maxEUOutput());
 			if (consumeInputs(recipe)) {
 				triggerMachine(true);
